Cohesive convergance problem

+++ MicroscaleModelling_randomSweep.py
@@ -3,22 +3,7 @@ from math import *
 import numpy as np
 import os
 import multiprocessing
-
 print'%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%\n', 'Multiscale Modelling, Microscale  \n',
-def hentePopulation():                 #Les fiber matrix populasjon
-    xy=list()
-    f = open(coordpath,'r')
-    tekst = f.read()
-    f.close()
-    lines = tekst.split('\n')
-    for line in lines:
-        data = line.split('\t')
-        a = float(data[0])  #   X
-        b = float(data[1])  #   Y
-        c = float(data[2])  #   R
-        xy.append([a,b,c])  #   lagre til liste
-    print 'Antall fiber = ',int(nf),'\tAntall fiberkoordinater = '+str(len(xy))
-    return xy
 def CreateNewRVEModel():
     # Creates RVE model and orphanmesh. Lage 2D RVE shell, meshe RVE, extrudere til 3D part, lage orphanmesh og sette cohesive elementtype paa Interface
     execfile(GitHub+Abaqus+'RVEsketching.py')           # Lage 2D RVE fra fiberpopulasjon data
@@ -28,9 +13,8 @@ def CreateNewRVEModel():
     execfile(GitHub+Abaqus+'RVEelementsets.py')         # Fiber, sizing og matrix elementer i set og Fiber center datums for material orientation
     execfile(GitHub + Abaqus + 'RVEproperties.py')                         # Sett materialegenskaper for elementset
     execfile(GitHub + Abaqus + 'RVE_Assembly_RP_CE.py')     # Assembly med RVE med x i fiber retning. Lage constrain equations til RVE modell og fixe boundary condition for rigid body movement
-    if not noFiber and Interface and False:        # Rearrange fiber interface nodes for controlled elementthickness and stable simulations
+    if not noFiber and Interface:        # Rearrange fiber interface nodes for controlled elementthickness and stable simulations
         execfile(GitHub + Abaqus + 'RVE_InterfaceElementThickness.py')
-
 def run_Job(Jobb, modelName):
     mdb.Job(name=Jobb, model=modelName, description='', type=ANALYSIS,
             atTime=None, waitMinutes=0, waitHours=0, queue=None, memory=90,
@@ -39,24 +23,6 @@ def run_Job(Jobb, modelName):
             modelPrint=OFF, contactPrint=OFF, historyPrint=OFF, userSubroutine='',
             scratch='', resultsFormat=ODB, multiprocessingMode=DEFAULT, numCpus=1,
             numDomains=1, numGPUs=1)
-    """mdb.Job(name=Jobb, type=ANALYSIS, explicitPrecision=SINGLE, nodalOutputPrecision=SINGLE,
-        parallelizationMethodExplicit=DOMAIN, numDomains=1,
-        model=modelName, description='',
-        atTime=None, queue=None,
-        memoryUnits=PERCENTAGE, getMemoryFromAnalysis=True,
-        echoPrint=OFF,
-        modelPrint=OFF, contactPrint=OFF, historyPrint=OFF, userSubroutine='',
-        scratch='', resultsFormat=ODB,
-        numGPUs=1)
-    
-    mdb.Job(name=Jobb, type=ANALYSIS,numCpus=numCPU, memory=95,
-            explicitPrecision=SINGLE, nodalOutputPrecision=SINGLE,
-            parallelizationMethodExplicit=DOMAIN,numDomains=numCPU,
-            model=modelName, description='', getMemoryFromAnalysis=True,
-            echoPrint=OFF, modelPrint=OFF, contactPrint=OFF, historyPrint=OFF,
-            scratch='', resultsFormat=ODB, multiprocessingMode=DEFAULT,
-            numGPUs=1)
-    """
     if Runjobs:
         mdb.jobs[Jobb].submit(consistencyChecking=OFF)
         mdb.jobs[Jobb].waitForCompletion()
@@ -66,27 +32,27 @@ def run_Job(Jobb, modelName):
         qw.write('call "C:\SIMULIA\Abaqus\6.14-4\code\bin\abq6144.exe" job=' + Jobb + ' cpus=' + str(numCPU))
         qw.close()
 numCPU = multiprocessing.cpu_count()
-Retning =['Exx', 'Eyy' , 'Ezz' ,'Exy' , 'Exz' , 'Eyz']
 
 """         PROCESS FLAGS                                       """
-
+RunningCleanup = 0
 Createmodel = 1
-Savemodel = 1
+Savemodel = 0
 
 Runjobs = 1                             #   ON/OFF Start analyser or create .inp
-
 linearAnalysis = 0                      #   ON/OFF Linear analyse for stiffness
-nonLinearDeformation = 1                #   ON/OFF non-linear analyse for strength
-
-Dampening = 0
+nonLinearAnalysis = 1                   #   ON/OFF non-linear analyse for strength
 
 Singlepin = 1                               #   Randbetingelse:    Laaser hjornenode mot forskyvning i 3 retninger
-tripplepin = 1                              #   Randbetingelse:    Laaser to noder mot forskyvning. En sentrert kantnode i 2 retninger og midtnode i 1 retning
-"""         RVE MODELLERING                """
+tripplepin = 0                              #   Randbetingelse:    Laaser to noder mot forskyvning. En sentrert kantnode i 2 retninger og midtnode i 1 retning
+
+Dampening = 1
+Stabl_Magn =2e-4
+Atapt_Damp_Ratio = 0.05
+Increments = {'maxNum': 1000, 'initial': 1e-3, 'min': 1e-5, 'max': 1e-1}
 
 Interface = 1                                   # ON/OFF CohesiveInterface
 rinterface = 0.001                              # Interfacetykkelse ved modellering. Verdi er relativ til radius.    0.01 = 1%
-ElementInterfaceT = 0.001                       # Interfacetykkelse paa elementene.  Verdi er relativ til radius.
+ElementInterfaceT = 0.1                       # Interfacetykkelse paa elementene.  Verdi er relativ til radius.
 
 noFibertest = 0                                     # ON/OFF Fiber i modellen.
 Fibervariation = 1                                  # ON/OFF variasjon fiberradius. Mean and standard div. Kan paavirke Vf i endelig model.
@@ -95,11 +61,11 @@ rmean = 8.7096                              # Gjennomsnittradius pa fiber
 Rstdiv = 0.6374                             # OStandard avvik fra gjennomsnittsradius
 
 # Meshsize
-FiberSirkelResolution = 24                                  # Meshresolution pa Fiber omkrets. 2*pi/FiberSirkelResolution
+FiberSirkelResolution = 20                                  # Meshresolution pa Fiber omkrets. 2*pi/FiberSirkelResolution
 meshsize = rmean * 2 * pi / FiberSirkelResolution           # Meshsize fra resolution paa interface paa fiberomkrets
 
 #Material Density
-MaterialDens  = 0
+MaterialDens  = 1
 
 
 Sample=[20]   #Forste sweepvariabel
@@ -111,7 +77,7 @@ for m in range(0,len(Sample)):
         Vf   =      0.6                 #
         RVEt =      meshsize            #      RVE tykkelse i fiberretning
 
-        Rclearing  = 0.025                    # Minimumsavstand mellom fiberkant og RVE kant. Verdi relativ til radius. Skal den settes lik meshsize?
+        Rclearing  = 0.05                    # Minimumsavstand mellom fiberkant og RVE kant. Verdi relativ til radius. Skal den settes lik meshsize?
         tol = rinterface*0.4                  # Modelleringstoleranse - Mindre en minste modelleringsvariabel (rInterface)
 
     """   Stess sweeps settings     """
@@ -133,7 +99,7 @@ for m in range(0,len(Sample)):
             #RVE tykkelse
         tykkelse = RVEt
 
-        """     Utregning av modelleringsvariabler   """
+        """     Modelleringsvariabler   """
         r = rmean
         gtol = Rclearing * r  # Dodsone klaring toleranse
         ytredodgrense = r + gtol  # Dodzone avstand, lengst fra kantene
@@ -142,7 +108,7 @@ for m in range(0,len(Sample)):
         iterasjonsgrense = 10000   # iterasjonsgrense i Fiberutplassering loop
 
         sweepresolution = 2 * pi / sweepcases
-
+        Retning = ['Exx', 'Eyy', 'Ezz', 'Exy', 'Exz', 'Eyz']
         """   Details  """
         if Interface and Createmodel:
             print('Aspect ratio for Interface elements ved modellering = ' + str(round(meshsize / (rinterface * rmean), 2)) +
@@ -154,8 +120,9 @@ for m in range(0,len(Sample)):
         Tekstfiler, Abaqus = 'textfiles/', 'Abaqus_steg/'                   #Undermapper
 
         Jobsss = workpath + 'Abaqusjobs.bat'
-        parameterpath = GitHub + 'Parametere.txt'               # Skrives ned for chaining eller importering
-        coordpath = GitHub +  'RVEcoordinater.txt'              # Skrives ned i genererefiberPop for chaining til main script
+        parameterpath = GitHub +Tekstfiler+'Parametere.txt'               # Skrives ned for chaining eller importering
+        iterationsParaPath = GitHub + Tekstfiler+'IterationParameters.txt'
+
 
         """ ABAQUS modelleringsnavn    """
 
@@ -191,13 +158,14 @@ for m in range(0,len(Sample)):
         import connectorBehavior
 
         # Rydde for neste iterasjon
-        """filelist = [f for f in os.listdir(workpath) if not f.endswith('.inp')]
-        for f in filelist:
-            try:
-                os.remove(os.path.join(workpath, f))
-            except:
-                pass
-        """
+        if RunningCleanup:
+            filelist = [f for f in os.listdir(workpath) if not f.endswith('.inp')]
+            for f in filelist:
+                try:
+                    os.remove(os.path.join(workpath, f))
+                except:
+                    pass
+
         """Random variabler og iterasjonsnavn"""
         if True:                     # For aa kunne kollapse variabler
             seed(Q)                                                         # Q er randomfunksjonensnokkelen
@@ -215,7 +183,7 @@ for m in range(0,len(Sample)):
             """ Datalagring """
             lagrestiffpath = GitHub + Tekstfiler + 'Stiffness__NF-'+ str(int(nf))+'.txt'  # Skrives ned statistikk til ett annet script
             Envelope = GitHub + Tekstfiler + 'envelope'  # Parameteravhengig - Spesifikt navn legges til i funksjonen
-
+            coordpath = GitHub + Tekstfiler + 'RVEcoordinatsandRadiuses'+ str(int(Sample[m])) + '_' + str(Q)+'.txt'  # Skrives ned i genererefiberPop for reference
         """ Get Abaqus RVE model """
         if True:
             Mdb()  # reset Abaqus
@@ -225,21 +193,16 @@ for m in range(0,len(Sample)):
                 xydata = None                       # Fiber kordinater og radiuser
                 if not noFiber:
                     execfile(GitHub+'GenerereFiberPopTilFil.py')            # create a random population
-                    xydata= hentePopulation()                               # hente fibercoordinater
                 CreateNewRVEModel()
                 if Savemodel:
                     if Interface:
+                        ### DEBUGGING FOR INTERFACE PROBLEMS
                         mdb.saveAs(pathName=workpath+'RVE-'+str(Sample[m])+'-'+str(int(Q)))
                     else:
                         mdb.saveAs(pathName=workpath + 'RVE-' + str(Sample[m]) + '-0-int-' + str(int(Q)))
             else:
-                if Interface:
-                    openMdb(pathName=workpath + 'RVE-' + str(Sample[m]) + '-' + str(int(Q)))
-                else:
-                    openMdb(pathName=workpath + 'RVE-' + str(Sample[m]) + '-0-int-' + str(int(Q)))
-                openMdb(pathName=workpath+'RVE-'+str(Sample[m])+'-'+str(int(Q)))
-                mod = mdb.models[modelName]
-        session.setValues(kernelMemoryLimit= 80000000)
+                openMdb(pathName=workpath + 'RVE-' + str(Sample[m]) + '-' + str(int(Q)))
+
         """ Boundaryconditions mot rigid body movement"""
         if Singlepin:
             region = mod.rootAssembly.sets['NL1']
@@ -266,11 +229,14 @@ for m in range(0,len(Sample)):
             except:
                 pass
                 n=n+1
-        if nonLinearDeformation:                            # nonLinearAnalysis for strength and large deformation
-            strain = 0.2#       STRAINS:  exx, eyy, ezz, exy, exz, eyz
-            strains = {'ShearExy': [0, round((-strain/6.0),4), 0, round((strain/2.0),4), 0, 0], 'TensionEyy': [0, 0.1, 0, 0, 0, 0], 'TensionEzz': [0, 0, 0.1, 0, 0, 0]}
+        if nonLinearAnalysis:                            # nonLinearAnalysis for strength and large deformation
+            strain = 0.03#       STRAINS:  exx, eyy, ezz, exy, exz, eyz
+            strains = {'ShearExy': [0, -strain/3, 0, strain, 0, 0], 'TensionEyy': [0, 0.1, 0, 0, 0, 0], 'TensionEzz': [0, 0, 0.1, 0, 0, 0]}
+            strains = {'ShearExy': [0, strain, 0, 0, 0, 0], 'TensionEyy': [0, 0.1, 0, 0, 0, 0], 'TensionEzz': [0, 0, 0.1, 0, 0, 0]}
+            #strains = {'ShearExy': [-strain,0,  0, 0, 0, 0], 'TensionEyy': [0, 0.1, 0, 0, 0, 0], 'TensionEzz': [0, 0, 0.1, 0, 0, 0]}
 
             if Interface:
+                ### DEBUGGING FOR INTERFACE PROBLEMS
                 #       CASES: Name, Strains
                 cases = [['ShearExy', strains['ShearExy']]]  # , ['TensionEyy',strains['TensionEyy']], ['TensionEzz',strains['TensionEzz']]]       # Shear + Compression
             else:
@@ -279,19 +245,31 @@ for m in range(0,len(Sample)):
             for Case in cases:
                 sim = 0
                 Jobbnavn, Strain = Case
+
                 try:
-                    Increments = {'maxNum': 1000, 'initial': 1e-8, 'min': 1e-12, 'max': 1e-2}
                     execfile(GitHub + Abaqus + 'nonLinearAnalysis.py')
                     sim=1
                 except:
                     pass
                 if not sim:
                     try:
-                        Increments = {'maxNum': 1000, 'initial': 1e-9, 'min': 1e-12, 'max': 1e-2}
+                        Increments = {'maxNum': 100, 'initial': 1e-4, 'min': 1e-10, 'max': 1e-1}
                         execfile(GitHub + Abaqus + 'nonLinearAnalysis.py')
                         sim = 1
                     except:
                         pass
+
+        try:
+            session.odbs['C:/Temp/ShearExyjob0.odb'].close()
+        except:
+            pass
+        o3 = session.openOdb(name='C:/Temp/ShearExyjob0.odb')
+        session.viewports['Viewport: 1'].setValues(displayedObject=o3)
+        session.viewports['Viewport: 1'].odbDisplay.display.setValues(plotState=(
+            CONTOURS_ON_UNDEF, CONTOURS_ON_DEF,))
+        leaf = dgo.LeafFromElementSets(elementSets=('PART-1-MESH-1-1.INTERFACES',))
+        session.viewports['Viewport: 1'].odbDisplay.displayGroup.replace(leaf=leaf)
+
         print 'Reached end of random key Iteration'
         Q = Q + 1
         del section, regionToolset, dgm, part, material, assembly, step, interaction
